// src/app/models/interfaces/configInterface.ts
import {configurations} from "./configurations";

export interface ConfigInterface {
  theme?: string;
  logoPath?: string;
  wishHeart?: string;
  filter?: configurations;
  behaviour?: configurations;
  name?: string;
  twitter?: string;
  instagram?: string;
  ligaPokemon?: string;
  description?: string;
}